fin_agent_hierarchical/requirements.txt
google-genai
langgraph
langchain
langchain-core
requests
python-dotenv
termcolor
pydantic
